requirements.txt
spidev
RPi.GPIO